+++ Apps/Admin/Client/Components/Dashboard/RatingSummary.razor
