+++ Apps/Admin/Client/Layouts/MainLayout.razor
@@ -1,3 +1,4 @@
+@using HealthGateway.Admin.Client.Components.Site
 @inherits LayoutComponentBase
 
 <MudThemeProvider Theme="@CurrentTheme" />
