+++ Apps/Admin/Client/Pages/CommunicationsPage.razor
@@ -3,6 +3,7 @@
 @attribute [Authorize(Roles = $"{Roles.Admin},{Roles.Reviewer}")]
 @using HealthGateway.Admin.Client.Store.Broadcasts
 @using HealthGateway.Admin.Client.Store.Communications
+@using HealthGateway.Admin.Client.Components.Communications
 @inherits Fluxor.Blazor.Web.Components.FluxorComponent
 
 <PageTitle>Health Gateway Admin Communications</PageTitle>
