+++ Apps/Admin/Client/Pages/CommunicationsPage.razor.cs
@@ -21,7 +21,8 @@ using System.Linq;
 using System.Threading.Tasks;
 using Fluxor;
 using Fluxor.Blazor.Web.Components;
-using HealthGateway.Admin.Client.Components;
+using HealthGateway.Admin.Client.Components.Common;
+using HealthGateway.Admin.Client.Components.Communications;
 using HealthGateway.Admin.Client.Models;
 using HealthGateway.Admin.Client.Store.Broadcasts;
 using HealthGateway.Admin.Client.Store.Communications;
