+++ Apps/Admin/Client/Pages/SupportPage.razor
@@ -6,6 +6,7 @@
 @using HealthGateway.Common.Data.Constants
 @using HealthGateway.Common.Data.Utils
 @using HealthGateway.Common.Ui.Constants
+@using HealthGateway.Admin.Client.Components.Support
 @inherits Fluxor.Blazor.Web.Components.FluxorComponent
 
 <PageTitle>Health Gateway Admin Support</PageTitle>
