+++ Apps/Admin/Client/_Imports.razor
@@ -11,7 +11,7 @@
 @using MudBlazor
 @using HealthGateway.Admin.Client
 @using HealthGateway.Admin.Client.Authorization
-@using HealthGateway.Admin.Client.Components
+@using HealthGateway.Admin.Client.Components.Common
 @using HealthGateway.Admin.Client.Layouts
 @using HealthGateway.Admin.Client.Store
 @using HealthGateway.Admin.Client.Utils
